Moving config and data folders to infrastructure layer
