+++ src/application/dtos/genre/genre-id.dto.ts
@@ -1,5 +1,5 @@
 import { ValidationResult } from '../../../domain';
-import { genreIdSchema, ZodAdapter } from '../../../config';
+import { genreIdSchema, ZodAdapter } from '../../../infrastructure';
 
 export class GenreIdDto {
   private constructor(public readonly genre_id: number) {}
