+++ src/application/dtos/review/update-review.dto.ts
@@ -1,5 +1,5 @@
 import { ValidationResult } from '../../../domain';
-import { updateReviewSchema, ZodAdapter } from '../../../config';
+import { updateReviewSchema, ZodAdapter } from '../../../infrastructure';
 
 export class UpdateReviewDto {
   private constructor(
