--- src/config/index.ts
@@ -1,14 +1,0 @@
-export * from './adapters/envs.adapter';
-export * from './adapters/jwt.adapter';
-export * from './adapters/zod.adapter';
-export * from './schemas/auth/auth-dtos.schemas';
-export * from './schemas/genre/genre-dtos.schemas';
-export * from './schemas/genre/genre.schema';
-export * from './schemas/movie/movie-dtos.schemas';
-export * from './schemas/movie/movie.schema';
-export * from './schemas/review/review-dtos.schemas';
-export * from './schemas/review/review.schema';
-export * from './schemas/shared/filters.schemas';
-export * from './schemas/shared/shared.schemas';
-export * from './schemas/user/user-dtos.schemas';
-export * from './schemas/user/user.schema';
