--- src/data/index.ts
@@ -1,2 +1,0 @@
-export * from './postgres/postgres-database';
-export * from './seed/data';
