+++ src/infrastructure/config/schemas/movie/movie-dtos.schemas.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 
-import { CreateMovieDto, MovieIdDto, UpdateMovieDto } from '../../../application';
-import { idSchema } from '../..';
+import { CreateMovieDto, MovieIdDto, UpdateMovieDto } from '../../../../application';
+import { idSchema } from '../../..';
 
 const baseMovieDtoSchema = z.object({
   title: z.string().min(2),
