+++ src/infrastructure/config/schemas/shared/filters.schemas.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 
-import { MovieFiltersDto } from '../../../application';
+import { MovieFiltersDto } from '../../../../application';
 
 export const movieFiltersSchema: z.ZodType<MovieFiltersDto> = z.object({
   title: z.string().min(2).optional(),
