+++ src/infrastructure/config/schemas/shared/shared.schemas.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 
-import { PaginationDto } from '../../../application';
+import { PaginationDto } from '../../../../application';
 
 export const paginationSchema: z.ZodType<PaginationDto> = z.object({
   page: z.number().positive().int(),
