+++ src/infrastructure/config/schemas/user/user-dtos.schemas.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 
-import { UpdateUserDto, UserIdDto } from '../../../application';
-import { idSchema } from '../..';
+import { UpdateUserDto, UserIdDto } from '../../../../application';
+import { idSchema } from '../../..';
 
 export const userIdSchema: z.ZodType<UserIdDto> = z.object({ user_id: z.lazy(() => idSchema) });
