+++ src/infrastructure/index.ts
@@ -1,3 +1,19 @@
+export * from './config/adapters/envs.adapter';
+export * from './config/adapters/jwt.adapter';
+export * from './config/adapters/zod.adapter';
+export * from './config/schemas/auth/auth-dtos.schemas';
+export * from './config/schemas/genre/genre-dtos.schemas';
+export * from './config/schemas/genre/genre.schema';
+export * from './config/schemas/movie/movie-dtos.schemas';
+export * from './config/schemas/movie/movie.schema';
+export * from './config/schemas/review/review-dtos.schemas';
+export * from './config/schemas/review/review.schema';
+export * from './config/schemas/shared/filters.schemas';
+export * from './config/schemas/shared/shared.schemas';
+export * from './config/schemas/user/user-dtos.schemas';
+export * from './config/schemas/user/user.schema';
+export * from './data/postgres/postgres-database';
+export * from './data/seed/data';
 export * from './datasources/auth.datasource.impl';
 export * from './datasources/genre.datasource.impl';
 export * from './datasources/movie.datasource.impl';
