+++ src/infrastructure/mappers/movie.mapper.ts
@@ -1,5 +1,5 @@
 import { CustomError, MovieEntity } from '../../domain';
-import { movieSchema, ZodAdapter } from '../../config';
+import { movieSchema, ZodAdapter } from '..';
 
 export class MovieMapper {
   static movieEntityFromObject(object: Record<string, any>): MovieEntity {
