+++ src/infrastructure/mappers/user.mapper.ts
@@ -1,5 +1,5 @@
 import { CustomError, UserEntity } from '../../domain';
-import { userSchema, ZodAdapter } from '../../config';
+import { userSchema, ZodAdapter } from '..';
 
 export class UserMapper {
   static userEntityFromObject(object: Record<string, any>): UserEntity {
